feat: seo notifications
