+++ apps/web/app/notifications/components/_Content.tsx
@@ -1,0 +1,106 @@
+'use client';
+
+import { Content, Typography } from '@social/ui-shared';
+import { useNotificationsContext, usePubkyClientContext } from '@/contexts';
+import {
+  CreatePost,
+  Header,
+  HotTags,
+  Sidebar,
+  Skeleton,
+  WhoFollow,
+} from '@/components';
+import Notifications from './_Notification';
+import Root from './_Root';
+import { useState, useEffect } from 'react';
+import { NotificationView } from '@/types/User';
+
+export default function Index() {
+  const { putTimestampNotification } = usePubkyClientContext();
+  const { notifications, loading } = useNotificationsContext();
+  const [loadingNotifications, setLoadingNotifications] = useState(true);
+  const timestamp = Date.now();
+
+  useEffect(() => {
+    const PutTimestamp = async () => {
+      await putTimestampNotification(timestamp);
+    };
+    PutTimestamp();
+  }, []);
+
+  useEffect(() => {
+    if (!loading) {
+      setLoadingNotifications(loading);
+    }
+  }, [loading]);
+
+  function renderNotifications(
+    notifications: NotificationView[],
+    loadingNotifications: boolean
+  ) {
+    if (loadingNotifications) {
+      return <Skeleton.Simple />;
+    }
+
+    if (notifications.length === 0) {
+      return (
+        <Typography.Body variant="small" className="text-opacity-50">
+          No notifications yet
+        </Typography.Body>
+      );
+    }
+
+    return notifications.map((notification, index) => {
+      {
+        /**if (Array.isArray(notification)) {
+        switch (notification[0].body.type) {
+          case 'Follow':
+          case 'NewFriend':
+          case 'LostFriend':
+            return (
+              <Notifications.NotificationGroup
+                key={index}
+                notifications={notification}
+              />
+            );
+          case 'TagProfile':
+            return (
+              <Notifications.NotificationTagGroup
+                key={index}
+                notifications={notification}
+              />
+            );
+          case 'TagPost':
+            return (
+              <Notifications.NotificationTagPostGroup
+                key={index}
+                notifications={notification}
+              />
+            );
+          default:
+            return null;
+        }
+      } else {
+        return (
+          <Notifications.Notification key={index} notification={notification} />
+        );
+      }*/
+      }
+      return <Notifications key={index} notification={notification} />;
+    });
+  }
+
+  return (
+    <Content.Main>
+      <Header className="hidden md:block" title="Notifications" />
+      <Content.Grid className="flex w-full justify-between items-start gap-12">
+        <Root>{renderNotifications(notifications, loadingNotifications)}</Root>
+        <Sidebar className="self-start sticky top-[120px] hidden xl:block w-[20%]">
+          <WhoFollow />
+          <HotTags />
+        </Sidebar>
+      </Content.Grid>
+      <CreatePost />
+    </Content.Main>
+  );
+}
