+++ apps/web/app/notifications/components/_NotificationGroup.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useState, useEffect } from 'react';
 import { Utils } from '@social/utils-shared';
 import { Icon, Typography, Button } from '@social/ui-shared';
